refactor: twitter entrypoint
