+++ skills/twitter/post_tweet.py
@@ -76,8 +76,8 @@ class TwitterPostTweet(TwitterBaseTool):
                     raise ValueError(
                         "Image upload is not supported when using API key authentication"
                     )
-                # Use the base class method to upload the image
-                media_ids = await self.upload_media(context.agent.id, image)
+                # Use the TwitterClient method to upload the image
+                media_ids = await twitter.upload_media(context.agent.id, image)
 
             # Post tweet using tweepy client
             tweet_params = {"text": text, "user_auth": twitter.use_key}
